+++ gadk/cli.py
@@ -1,6 +1,8 @@
+import abc
 import importlib
+import inspect
 from os import getcwd, makedirs
-from typing import Any, List
+from typing import Any, List, Set
 
 import click
 import sys
@@ -20,7 +22,7 @@ def output_to_stdout(workflow: Workflow):
     print(workflow.render())
 
 
-def workflows_from_module(actions: Any) -> List[Workflow]:
+def find_workflows() -> List[Workflow]:
     """
     Extract workflows from imported module.
 
@@ -29,12 +31,42 @@ def workflows_from_module(actions: Any) -> List[Workflow]:
         2. Workflow subclasses should define a constructor with no arguments. The arguments
            exist for the programmer to name the workflow. `gadk` cannot guess these arguments.
     """
-    subclasses = actions.Workflow.__subclasses__()  # type: ignore
-    workflows: List[Workflow] = [workflow() for workflow in subclasses]  # type: ignore
-    return workflows
+
+    def _find_workflows(subclasses: List, workflows: Set) -> Set:
+        """
+        Recursive function to find workflows by descending a class hierarchy of abstract workflows.
+
+        Simple workflows will return immediately. More complex projects might recurse once or twice.
+        """
+        child_workflows = []
+        for workflow_class in subclasses:
+            # Add subclasses of abstract subclasses. This allows for further abstractions of workflows.
+            if inspect.isabstract(workflow_class):
+                child_workflows += [
+                    child_workflow
+                    for child_workflow in workflow_class.__subclasses__()
+                    if child_workflow not in workflows
+                ]
+            else:
+                workflows.add(workflow_class)
+
+        if child_workflows:
+            return _find_workflows(child_workflows, workflows)
+        return workflows
+
+    # Collect the workflows detected in the module. There may be abstractions of workflows,
+    # so we'll look at subclasses of subclasses, and so on, if necessary.
+    workflows = _find_workflows(Workflow.__subclasses__(), set())
+
+    # Filter out those abstract workflows. Only concrete workflows should be returned.
+    return [
+        workflow_class()
+        for workflow_class in workflows
+        if not inspect.isabstract(workflow_class)
+    ]
 
 
-@click.command(context_settings={"help_option_names": ["-h", "--help"]},)
+@click.command(context_settings={"help_option_names": ["-h", "--help"]})
 @click.option(
     "--print/--no-print",
     default=False,
@@ -45,14 +77,16 @@ def cmd(print: bool):
 
     # Import actions.py from the current working directory.
     sys.path.insert(0, getcwd())
-    actions = importlib.import_module("actions")
+    importlib.import_module("actions")
     sys.path.pop(0)
 
     # Determine output per workflow.
     outputter = output_to_stdout if print else output_to_file
 
     # Assume actions.py imports all elements of gadk to get subclasses of Workflow.
-    for workflow in workflows_from_module(actions):
+    # Sort workflows for consistency.
+    workflows = sorted(find_workflows(), key=lambda w: w.name)
+    for workflow in workflows:
         outputter(workflow)