+++ tests/integration/examples/abstract/workflows/__init__.py
@@ -1,0 +1,2 @@
+from .foo import FooService
+from .bar import BarService