+++ tests/integration/examples/abstract/workflows/foo.py
@@ -1,0 +1,9 @@
+from .lib import Service
+
+
+class FooService(Service):
+    def __init__(self) -> None:
+        super().__init__("foo")
+
+    def service_name(self) -> str:
+        return "foo"