+++ tests/integration/tests.sh
@@ -8,6 +8,11 @@
 
 set -e
 
+# Make filenames specific to Python version. This way the output of tox tests is separated.
+python_version=$(python --version | tail -c +8)
+actual_filename="actual-$python_version.yml"
+errors_filename="errors-$python_version.log"
+diff_filename="diff-$python_version.log"
 
 passed=0
 
@@ -16,23 +21,23 @@ do
   cd "$test_scenario"
 
   set +e
-  gadk --print 1> actual.yml 2> errors.log
+  gadk --print 1> $actual_filename 2> $errors_filename
   set -e
 
   # Confirm that there are no errors.
-  if [[ $(wc -l errors.log) == "0 errors.log" ]]
+  if [[ $(wc -l $errors_filename) == "0 $errors_filename" ]]
   then
     echo "No errors in $test_scenario"
   else
     echo "Errors printed for $test_scenario"
-    cat errors.log
+    cat $errors_filename
     passed=1
     cd - > /dev/null
     continue
   fi
 
   # Confirm that actual.yml is not empty.
-  if [[ $(wc -l actual.yml) == "0 actual.yml" ]]
+  if [[ $(wc -l $actual_filename) == "0 $actual_filename" ]]
   then
     echo "$test_scenario seems to be WIP!"
     passed=1
@@ -40,8 +45,8 @@ do
     continue
   else
     # Confirm that actual.yml matches expected.yml.
-    diff expected.yml actual.yml > diff.log
-    if [[ $(wc -l diff.log) == "0 diff.log" ]]
+    diff expected.yml $actual_filename > $diff_filename
+    if [[ $(wc -l $diff_filename) == "0 $diff_filename" ]]
     then
       echo "$test_scenario has passed!"
     else