+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 0.1.1
+
+* Abstractions of workflows are now supported:
+  ```python
+  class BaseService(Workflow, ABC):
+     @abstractmethod
+     def concrete_data(self):
+        pass
+
+  class ServiceA(BaseService):
+    pass
+  ```
+  In the example above, `ServiceA` will be detected as a workflow.
+  `BaseService` will not be rendered as expected; only concrete workflows
+  are rendered.
+
 # 0.1.0
 
 * Initial release.